task_1_ai_fire_prediction_platform/ai_fire_prediction_requirements.txt
# Core dependencies for Synthetic Fire Prediction System
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.9.0
torchvision>=0.10.0
xgboost>=1.5.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
opencv-python>=4.5.0
Pillow>=8.3.0
PyYAML>=5.4.0
jsonschema>=3.2.0
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910
mlflow>=1.20.0
dvc>=2.6.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5
aiofiles>=0.7.0
asyncio-mqtt>=0.10.0
psutil>=5.8.0
boto3>=1.26.0
streamlit>=1.27.0